Replace old search component with new one, drop gridjs
