+++ src/components/Search.svelte
@@ -1,13 +1,10 @@
 <style lang="scss">
-  @import "/public/styles/mermaid.min.css";
-
   #app {
-    min-height: 100vh;
     display: flex;
     flex-direction: column;
     overflow: hidden;
-    margin: 10px;
-    gap: 0.5rem;
+    margin: 1rem;
+    gap: 1rem;
 
     .search-box {
       height: 2.25em;
@@ -80,103 +77,152 @@
     }
   }
 
-  :global(.row-links a) {
+  .row-links a {
     color: var(--primary-accent);
-    margin-left: 10px;
-    margin-right: 10px;
     display: inline-block;
     height: 20px;
+
+    :global(svg) {
+      height: 24px;
+      width: 24px;
+    }
+  }
+
+  /************************/
+  table {
+    border-collapse: collapse;
+    box-shadow:
+      0 1px 3px 0 rgba(0, 0, 0, 0.1),
+      0 1px 2px 0 rgba(0, 0, 0, 0.26);
+    margin: 0;
+    padding: 0;
+    table-layout: fixed;
+    text-align: left;
+    width: 100%;
+
+    th {
+      background-color: #f9fafb;
+      border: 1px solid #e5e7eb;
+      box-sizing: border-box;
+      color: #6b7280;
+      outline: none;
+      overflow: hidden;
+      padding: 14px 24px;
+      position: relative;
+      text-overflow: ellipsis;
+      user-select: none;
+      vertical-align: middle;
+      white-space: nowrap;
+
+      &:not(:first-child) {
+        background-image: url(data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIHdpZHRoPSI0MDIiIGhlaWdodD0iNDAyIiBzdHlsZT0iZmlsbDojMDAwNCI+PHBhdGggZD0iTTczLjA5MiAxNjQuNDUyaDI1NS44MTNjNC45NDkgMCA5LjIzMy0xLjgwNyAxMi44NDgtNS40MjQgMy42MTMtMy42MTYgNS40MjctNy44OTggNS40MjctMTIuODQ3cy0xLjgxMy05LjIyOS01LjQyNy0xMi44NUwyMTMuODQ2IDUuNDI0QzIxMC4yMzIgMS44MTIgMjA1Ljk1MSAwIDIwMC45OTkgMHMtOS4yMzMgMS44MTItMTIuODUgNS40MjRMNjAuMjQyIDEzMy4zMzFjLTMuNjE3IDMuNjE3LTUuNDI0IDcuOTAxLTUuNDI0IDEyLjg1IDAgNC45NDggMS44MDcgOS4yMzEgNS40MjQgMTIuODQ3IDMuNjIxIDMuNjE3IDcuOTAyIDUuNDI0IDEyLjg1IDUuNDI0ek0zMjguOTA1IDIzNy41NDlINzMuMDkyYy00Ljk1MiAwLTkuMjMzIDEuODA4LTEyLjg1IDUuNDIxLTMuNjE3IDMuNjE3LTUuNDI0IDcuODk4LTUuNDI0IDEyLjg0N3MxLjgwNyA5LjIzMyA1LjQyNCAxMi44NDhMMTg4LjE0OSAzOTYuNTdjMy42MjEgMy42MTcgNy45MDIgNS40MjggMTIuODUgNS40MjhzOS4yMzMtMS44MTEgMTIuODQ3LTUuNDI4bDEyNy45MDctMTI3LjkwNmMzLjYxMy0zLjYxNCA1LjQyNy03Ljg5OCA1LjQyNy0xMi44NDggMC00Ljk0OC0xLjgxMy05LjIyOS01LjQyNy0xMi44NDctMy42MTQtMy42MTYtNy44OTktNS40Mi0xMi44NDgtNS40MnoiLz48L3N2Zz4=);
+        background-position-x: calc(100% - 9px);
+        background-position-y: 50%;
+        background-repeat: no-repeat;
+        background-size: 12px;
+        cursor: pointer;
+
+        &.sortedAsc {
+          background-image: url(data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIHdpZHRoPSIyOTIuMzYyIiBoZWlnaHQ9IjI5Mi4zNjEiIHN0eWxlPSJlbmFibGUtYmFja2dyb3VuZDpuZXcgMCAwIDI5Mi4zNjIgMjkyLjM2MSIgeG1sOnNwYWNlPSJwcmVzZXJ2ZSI+PHBhdGggZD0iTTI4Ni45MzUgMTk3LjI4NyAxNTkuMDI4IDY5LjM4MWMtMy42MTMtMy42MTctNy44OTUtNS40MjQtMTIuODQ3LTUuNDI0cy05LjIzMyAxLjgwNy0xMi44NSA1LjQyNEw1LjQyNCAxOTcuMjg3QzEuODA3IDIwMC45MDQgMCAyMDUuMTg2IDAgMjEwLjEzNHMxLjgwNyA5LjIzMyA1LjQyNCAxMi44NDdjMy42MjEgMy42MTcgNy45MDIgNS40MjUgMTIuODUgNS40MjVoMjU1LjgxM2M0Ljk0OSAwIDkuMjMzLTEuODA4IDEyLjg0OC01LjQyNSAzLjYxMy0zLjYxMyA1LjQyNy03Ljg5OCA1LjQyNy0xMi44NDdzLTEuODE0LTkuMjMtNS40MjctMTIuODQ3eiIvPjwvc3ZnPg==);
+          background-position-x: calc(100% - 10px);
+          background-position-y: 45%;
+          background-size: 10px;
+        }
+
+        &.sortedDesc {
+          background-image: url(data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIHdpZHRoPSIyOTIuMzYyIiBoZWlnaHQ9IjI5Mi4zNjIiIHN0eWxlPSJlbmFibGUtYmFja2dyb3VuZDpuZXcgMCAwIDI5Mi4zNjIgMjkyLjM2MiIgeG1sOnNwYWNlPSJwcmVzZXJ2ZSI+PHBhdGggZD0iTTI4Ni45MzUgNjkuMzc3Yy0zLjYxNC0zLjYxNy03Ljg5OC01LjQyNC0xMi44NDgtNS40MjRIMTguMjc0Yy00Ljk1MiAwLTkuMjMzIDEuODA3LTEyLjg1IDUuNDI0QzEuODA3IDcyLjk5OCAwIDc3LjI3OSAwIDgyLjIyOGMwIDQuOTQ4IDEuODA3IDkuMjI5IDUuNDI0IDEyLjg0N2wxMjcuOTA3IDEyNy45MDdjMy42MjEgMy42MTcgNy45MDIgNS40MjggMTIuODUgNS40MjhzOS4yMzMtMS44MTEgMTIuODQ3LTUuNDI4TDI4Ni45MzUgOTUuMDc0YzMuNjEzLTMuNjE3IDUuNDI3LTcuODk4IDUuNDI3LTEyLjg0NyAwLTQuOTQ4LTEuODE0LTkuMjI5LTUuNDI3LTEyLjg1eiIvPjwvc3ZnPg==);
+          background-position-x: calc(100% - 10px);
+          background-position-y: 55%;
+          background-size: 10px;
+        }
+
+        &:hover,
+        &:focus {
+          background-color: #e5e7eb;
+        }
+      }
+    }
+
+    th:first-child,
+    td:first-child,
+    th:last-child,
+    td:last-child {
+      width: min(16vw, 240px);
+    }
+
+    tbody {
+      background-color: #fff;
+    }
+
+    td {
+      border: 1px solid #e5e7eb;
+      box-sizing: content-box;
+      height: 64px;
+      padding: 0 min(1.5%, 24px);
+
+      &:first-child div {
+        display: flex;
+        align-items: center;
+        justify-content: space-evenly;
+        gap: 0.3rem;
+        flex-wrap: wrap;
+      }
+    }
+
+    tfoot {
+      background-color: #f9fafb;
+      border: 1px solid #e5e7eb;
+
+      div {
+        align-items: center;
+        display: flex;
+
+        span {
+          flex-grow: 1;
+        }
+      }
+
+      button {
+        background-color: #fff;
+        border: 1px solid #d2d6dc;
+        cursor: pointer;
+        padding: 5px 14px;
+        user-select: none;
+
+        &[disabled] {
+          cursor: not-allowed;
+        }
+
+        &:not([disabled]):hover {
+          background-color: #f7f7f7;
+          color: #3c4257;
+        }
+      }
+    }
   }
 </style>
 
 <script>
-  import Grid from "gridjs-svelte";
-  import { html } from "gridjs";
+  import Icon from "../ui-components/Icon.svelte";
   import catalog from "../config/catalog.json";
 
-  let grid = null;
-  let input;
+  // Prep the composerArranger property
+  catalog.forEach((item) => {
+    item.composerArranger = [
+      ...new Set([item.composer, item.arranger].filter(Boolean)),
+    ].join("\n");
+    item.publisher = `${item.publisher}  [${item.number}]`;
+  });
 
-  let filteredListItems;
+  let filteredListItems = catalog;
   let activeFacet;
+  let sortOrder;
+  let searchParts = [];
+  let pageSize = 10;
+  let currentPage = 1;
+  let filteredAndPagedItems;
 
-  const searchFields = ["title", "composer_arranger", "performer", "publisher"];
-
-  // taken from Icon component
-  // The way grid.js loads the values (and icons/links) for each row as HTML
-  // makes it very difficult to reuse that component here.
-  const icons = {
-    piano: `
-      <svg height="24" width="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round" name="piano" role="img" aria-label="Perform Roll">
-        <rect x="2" y="7" width="20" height="12" rx="2"></rect>
-        <line x1="7" y1="9" x2="7" y2="13.5" style="stroke-width:2.5;stroke-linecap:square"></line>
-        <line x1="12" y1="9" x2="12" y2="13.5" style="stroke-width:2.5;stroke-linecap:square"></line>
-        <line x1="17" y1="9" x2="17" y2="13.5" style="stroke-width:2.5;stroke-linecap:square"></line>
-      </svg>
-      `,
-    play: `
-      <svg height="24" width="24" viewBox="0 0 24 24" stroke-width="2" stroke="currentColor" fill="none" stroke-linecap="round" stroke-linejoin="round" name="play" role="img" aria-label="Play Roll">
-        <path
-          fill-rule="evenodd"
-          clip-rule="evenodd"
-          d="M12 21C16.9706 21 21 16.9706 21 12C21 7.02944 16.9706 3 12 3C7.02944 3 3 7.02944 3 12C3 16.9706 7.02944 21 12 21ZM12 23C18.0751 23 23 18.0751 23 12C23 5.92487 18.0751 1 12 1C5.92487 1 1 5.92487 1 12C1 18.0751 5.92487 23 12 23Z"
-          fill="currentColor"
-        />
-        <path d="M16 12L10 16.3301V7.66987L16 12Z" fill="currentColor" />
-      </svg>
-      `,
-    midi: `
-      <svg height="24" width="24" viewBox="0 0 712 266" stroke-width="2" stroke="currentColor" fill="currentColor" stroke-linecap="round" stroke-linejoin="round" name="midi" role="img" aria-label="Download MIDI">
-        <path aria-label="M" text-data="M" d="m 0,0 h 233 c 19.6,0 31,16.9 31,37 V 266 H 199 V 67 H 161 V 266 H 102 V 67 H 65 V 266 H 0 Z" />
-        <rect aria-label="I" text-data="I" x="296" y="0" width="65" height="266" />
-        <path aria-label="D" text-data="D" d="m 392,0 h 193 c 19.6,0 31,16.9 31,37 v 196 c 0,24.9 -10.4,33 -33,33 H 392 V 97 h 66 v 104 h 93 V 60 H 392 Z" />
-        <rect aria-label="I" text-data="I" x="646" y="0" width="66" height="266" />
-      </svg>
-      `,
-    roll_image: `
-      <svg height="24" width="24" viewBox="0 0 300 550" stroke-width="2" stroke="currentColor" fill="currentColor" stroke-linecap="round" stroke-linejoin="round" name="roll-image" role="img" aria-label="Download Image">
-        <rect x="1" y="50" width="298" height="200" stroke-width="4"></rect>
-        <path fill="lightgrey" stroke-width="4" d="M 1 250 V 400 L 135 549 H 165 L 299 400 V 250 Z"></path>
-        <rect x="100" y="400" width="100" height="75" rx="5%" fill="white" stroke-width="2"></rect>
-      </svg>
-      `,
-  };
-
-  const getLinksForCell = (druid) => {
-    const catalogEntry = catalog.filter((obj) =>
-      Object.keys(obj).some((key) => obj[key].includes(druid)),
-    )[0];
-    const imageKey = catalogEntry.image_url.split("/").slice(-2, -1)[0];
-    const imageLink = `https://stacks.stanford.edu/file/${druid}/${imageKey}.jp2`;
-
-    return html(`
-      <span class="row-links">
-        <a href="/?druid=${druid}" title="Play roll" target="_blank">
-          ${icons.play}
-        </a>
-        <a href="/perform/?druid=${druid}" title="Perform roll" target="_blank">
-          ${icons.piano}
-        </a>
-        <a href="/midi/${druid}.mid" title="Download MIDI">
-          ${icons.midi}
-        </a>
-        <a href="${imageLink}" title="Download image">
-          ${icons.roll_image}
-        </a>
-      </span>
-    `);
-  };
-
-  const pagination = {
-    enabled: true,
-    limit: 20,
-  };
-
-  const facetFilter = (listItem) => listItem.type === activeFacet;
-
-  const activateInput = () => {
-    input.focus();
-  };
+  const searchFields = ["title", "composerArranger", "performer", "publisher"];
 
   const unDecomposableMap = {
     ł: "l",
@@ -206,74 +252,6 @@
       .replace(/[\u0300-\u036f]/g, "")
       .trim();
 
-  const listItems = catalog.map((item) => {
-    // _s_ prefix is for search ( normalizedText is called )
-    // _d_ prefix is for display (they can be marked up)
-    const listItem = {
-      druid: item.druid,
-      title: item.work,
-      composer_arranger: [
-        ...new Set([item.composer, item.arranger].filter(Boolean)),
-      ].join(" <br/>"),
-      performer: item.performer,
-      publisher: `${item.publisher} (${item.number})`,
-      type: item.type,
-      _d_links: getLinksForCell(item.druid),
-    };
-    const searchArr = [];
-    searchFields.forEach((k) => {
-      searchArr.push(listItem[k]);
-      listItem[`_s_${k}`] = normalizeText(listItem[k]);
-      listItem[`_d_${k}`] = html(listItem[k]);
-    });
-
-    return { ...listItem, _search: normalizeText(searchArr.join("   ")) };
-  });
-
-  const sortColumnHtml = (a, b) => {
-    if (a.props.content > b.props.content) {
-      return 1;
-    }
-    if (b.props.content > a.props.content) {
-      return -1;
-    }
-    return 0;
-  };
-
-  const columns = [
-    {
-      id: "_d_links",
-      name: "Play/Perform/MIDI/Image",
-      width: "240px",
-      sort: false,
-      data: (r) => r._d_links,
-    },
-    {
-      id: "_d_title",
-      name: "Title",
-      sort: { compare: sortColumnHtml },
-      data: (r) => r._d_title,
-    },
-    {
-      id: "_d_composer_arranger",
-      name: "Composer / Arranger",
-      sort: { compare: sortColumnHtml },
-      data: (r) => r._d_composer_arranger,
-    },
-    {
-      id: "_d_performer",
-      name: "Performer",
-      sort: { compare: sortColumnHtml },
-      data: (r) => r._d_performer,
-    },
-    {
-      id: "_d_publisher",
-      name: "Publisher",
-      sort: { compare: sortColumnHtml },
-      data: (r) => r._d_publisher,
-    },
-  ];
-
   const startIdxAdjustment = (str, idx) =>
     (str.toLowerCase().substring(0, idx).match(longSubstitutionsRegex) || [])
       .length;
@@ -283,10 +261,11 @@
       str.toLowerCase().substring(0, idx).match(longSubstitutionsRegex) || []
     ).reduce((adj, m) => adj + (unDecomposableMap[m].length - m.length), 0);
 
-  const markupMatches = (label, searchContent, searchParts) => {
+  const markupMatches = (label) => {
     const matchExtents = [];
     const mergedExtents = [];
     let markedUp = label;
+    const searchContent = normalizeText(label);
 
     searchParts.forEach((searchPart) => {
       let idx = -1;
@@ -319,57 +298,65 @@
           end,
         )}</mark>${markedUp.substring(end)}`;
       });
-    return html(markedUp);
+    return markedUp;
+  };
+
+  const prepSearchParts = (text) => {
+    searchParts = text
+      ? normalizeText(text.replace(/<br>|[&/\\#,+()$~%.'":*?<>{}]|nbsp;/g, " "))
+          .split(" ")
+          .slice(0, 8)
+      : [];
   };
 
   const itemFilter = async () => {
+    filteredListItems = catalog;
     if (activeFacet) {
-      filteredListItems = listItems.filter(facetFilter);
-    } else {
-      filteredListItems = listItems;
+      filteredListItems = filteredListItems.filter(
+        (listItem) => listItem.type === activeFacet,
+      );
     }
 
-    if (!input) return;
-    const filteredText = normalizeText(
-      input.value.replace(/<br>|[&/\\#,+()$~%.'":*?<>{}]|nbsp;/g, " "),
-    );
-
-    if (filteredText) {
-      const searchParts = filteredText.split(" ").slice(0, 8);
-      filteredListItems = filteredListItems
-        .filter((listItem) =>
-          searchParts.every((searchPart) =>
-            listItem._search.includes(searchPart),
+    if (searchParts.length) {
+      filteredListItems = filteredListItems.filter((item) =>
+        searchParts.some((searchPart) =>
+          searchFields.some((field) =>
+            normalizeText(item[field]).includes(searchPart),
           ),
-        )
-        .map((item) => {
-          searchFields.forEach(
-            (k) =>
-              (item[`_d_${k}`] = markupMatches(
-                item[k],
-                item[`_s_${k}`],
-                searchParts,
-              )),
-          );
-          return item;
-        });
-    } else {
-      // no search term so we need to be sure to remove any hit highlights that might still be present.
-      filteredListItems = filteredListItems.map((item) => {
-        searchFields.forEach((k) => (item[`_d_${k}`] = html(item[k])));
-        return item;
-      });
+        ),
+      );
     }
+    currentPage = 1;
+    paginate();
   };
 
-  const setActiveFacet = (facet) => {
-    activeFacet = facet === activeFacet ? undefined : facet;
+  const sortItems = (key) => {
+    sortOrder = sortOrder === `${key}-asc` ? `${key}-desc` : `${key}-asc`;
+    const asc = sortOrder.endsWith("asc");
+    filteredListItems = filteredListItems.sort((a, b) => {
+      if (a[key] < b[key]) return asc ? -1 : 1;
+      if (a[key] > b[key]) return asc ? 1 : -1;
+      return 0;
+    });
+    currentPage = 1;
+    paginate();
   };
 
-  const facets = [...new Set(listItems.map((item) => item.type))];
+  const setActiveFacet = (facet) =>
+    (activeFacet = facet === activeFacet ? undefined : facet);
+  const facets = [...new Set(catalog.map((item) => item.type))];
+
+  const paginate = () => {
+    filteredAndPagedItems = filteredListItems.slice(
+      0 + (currentPage - 1) * pageSize,
+      0 + (currentPage - 1) * pageSize + pageSize,
+    );
+  };
 
   /* eslint-disable no-unused-expressions, no-sequences */
   $: activeFacet, itemFilter();
+  $: searchParts, itemFilter();
+  $: currentPage, paginate();
 </script>
 
 <div id="app">
@@ -384,9 +371,7 @@
       contenteditable="true"
       aria-label="Search"
       aria-multiline="false"
-      bind:this={input}
-      on:focus={activateInput}
-      on:input={itemFilter}
+      on:input={(e) => prepSearchParts(e.target.value)}
     />
   </div>
   <!-- search-box -->
@@ -413,15 +398,125 @@
         {/each}
       </ul>
     {/if}
-    Filtered: {filteredListItems?.length} / {listItems.length}
+    Filtered: {filteredListItems?.length} / {catalog.length}
   </div>
   <!-- facets -->
-  <Grid
-    data={filteredListItems}
-    {columns}
-    {pagination}
-    resizable={true}
-    bind:this={grid}
-  />
+  <table>
+    <thead>
+      <tr>
+        <th>Play/Perform/MIDI/Image</th>
+        <th
+          tabindex="0"
+          on:click={() => sortItems("work")}
+          on:keypress={({ key }) => key === "Enter" && sortItems("work")}
+          class:sortedAsc={sortOrder === "work-asc"}
+          class:sortedDesc={sortOrder === "work-desc"}>Title</th
+        >
+        <th
+          tabindex="0"
+          on:click={() => sortItems("composerArranger")}
+          on:keypress={({ key }) =>
+            key === "Enter" && sortItems("composerArranger")}
+          class:sortedAsc={sortOrder === "composerArranger-asc"}
+          class:sortedDesc={sortOrder === "composerArranger-desc"}
+          >Composer / Arranger</th
+        >
+        <th
+          tabindex="0"
+          on:click={() => sortItems("performer")}
+          on:keypress={({ key }) => key === "Enter" && sortItems("performer")}
+          class:sortedAsc={sortOrder === "performer-asc"}
+          class:sortedDesc={sortOrder === "performer-desc"}>Performer</th
+        >
+        <th
+          tabindex="0"
+          on:click={() => sortItems("publisher")}
+          on:keypress={({ key }) => key === "Enter" && sortItems("publisher")}
+          class:sortedAsc={sortOrder === "publisher-asc"}
+          class:sortedDesc={sortOrder === "publisher-desc"}>Publisher</th
+        >
+      </tr>
+    </thead>
+    <tbody>
+      {#each filteredAndPagedItems as item}
+        {@const imageLink = `https://stacks.stanford.edu/file/${item.druid}/${item.image_url.split("/").slice(-2, -1)[0]}.jp2`}
+        <tr>
+          <td class="row-links">
+            <div>
+              <a href="/?druid={item.druid}" title="Play roll" target="_blank">
+                <Icon name="play" aria-label="Play roll" />
+              </a>
+              <a
+                href="/perform/?druid={item.druid}"
+                title="Perform roll"
+                target="_blank"
+              >
+                <Icon name="piano" aria-label="Perform roll" />
+              </a>
+              <a href="/midi/{item.druid}.mid" title="Download MIDI">
+                <Icon name="midi" aria-label="Download MIDI" />
+              </a>
+              <a href={imageLink} title="Download image">
+                <Icon name="roll-image" aria-label="Download Image" />
+              </a>
+            </div>
+          </td>
+          <td>
+            {@html searchParts.length ? markupMatches(item.work) : item.work}
+          </td>
+          <td>
+            {@html searchParts.length
+              ? markupMatches(item.composerArranger).replace("\n", "<br/>")
+              : item.composerArranger.replace("\n", "<br/>")}
+          </td>
+          <td>
+            {@html searchParts.length
+              ? markupMatches(item.performer)
+              : item.performer}
+          </td>
+          <td>
+            {@html searchParts.length
+              ? markupMatches(item.publisher)
+              : item.publisher}
+          </td>
+        </tr>
+      {/each}
+    </tbody>
+    <tfoot>
+      <tr>
+        <td colspan="5">
+          <div>
+            {#if filteredListItems.length === 0}
+              No results
+            {:else}
+              {@const pageStart = pageSize * (currentPage - 1) + 1}
+              <span>
+                Showing <strong>{pageStart}</strong> to
+                <strong>
+                  {Math.min(pageStart + pageSize - 1, filteredListItems.length)}
+                </strong>
+                of <strong>{filteredListItems.length}</strong>
+              </span>
+
+              <button
+                disabled={currentPage === 1}
+                on:click={() => currentPage--}
+                on:keypress={({ code }) => code === "Enter" && currentPage--}
+              >
+                &laquo; Previous
+              </button>
+              <button
+                disabled={currentPage * pageSize >= filteredListItems.length}
+                on:click={() => currentPage++}
+                on:keypress={({ code }) => code === "Enter" && currentPage++}
+              >
+                Next &raquo;
+              </button>
+            {/if}
+          </div>
+        </td>
+      </tr>
+    </tfoot>
+  </table>
 </div>
 <!-- app -->
