--- src/pages/search2/index.astro
@@ -1,15 +1,0 @@
----
-import BaseLayout from "../../layouts/base.astro";
-import Search from "../../components/Search2.svelte";
-import StanfordHeader from "../../layouts/stanfordHeader.astro";
-import StanfordSubheader from "../../layouts/stanfordSubheader.astro";
-// import NewHeader from "../../pages/newHeader.astro";
-import StanfordFooter from "../../layouts/stanfordFooter.astro";
----
-
-<BaseLayout title="Pianolatron Search">
-  <StanfordHeader />
-  <StanfordSubheader />
-  <Search client:only />
-  <StanfordFooter />
-</BaseLayout>
